refactor(RandomNumber.java):파일 위치 조정
RandomNumber.java파일을 /utils 내부로 이동

BREKAING CHANGE: RandomNumner.java 파일 위치 수정
as-is:
main/java/racingcar/RandomNumber.java
to-be:
main/java/racingcar/utils/RandomNumber.java
