+++ src/test/java/racingcar/utils/RandomNumberTest.java
@@ -1,10 +1,11 @@
-package racingcar;
+package racingcar.utils;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import racingcar.utils.RandomNumber;
 
 class RandomNumberTest {
 	@Test
